requirements.txt
# Core dependencies
playwright>=1.40.0
asyncio-mqtt>=0.13.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pyyaml>=6.0.1
aiohttp>=3.9.0
click>=8.1.7
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.0.0

# Optional currency conversion
requests>=2.31.0
